requirement.txt
pymongo==4.3.3
certifi
dnspython>=1.13.0,<2.0.0
python-dotenv==1.0.0
pyOpenSSL==23.0.0
pandas
numpy
pymongo[srv]==4.3.3
scikit-learn
dill
pyyaml
fastapi
uvicorn

-e .
